spaceVogageV0.13
UI Establish coded, progress and player health token basic

+++ SpacePirates/Assets/Scipts/Movements/SlimeEmeny.cs
@@ -6,8 +6,8 @@ public class SlimeEmeny : BaseMovement
     float movementHoldClock = 0;
 
     [Header("Damage")]
-    [SerializeField] float playerDamage = 1;
-    [SerializeField] float breakableDamage = 1;
+    [SerializeField] int playerDamage = 1;
+    [SerializeField] int breakableDamage = 1;
 
     bool damage;
